test: simplify test running
