+++ .gitignore
@@ -48,6 +48,7 @@ coverage.xml
 
 # Django stuff:
 *.log
+simple_test_db
 
 # Sphinx documentation
 docs/_build/
