+++ stream_django/tests/test_app/models.py
@@ -1,5 +1,6 @@
-from django.db import models
 from django.conf import settings
+from django.db import models
+
 from stream_django.activity import Activity
