--- stream_django/tests/test_app/tests/__init__.py
@@ -1,5 +1,0 @@
-# needed for Django < 1.6
-# flake8: noqa
-from .test_activity import *
-from .test_manager import *
-from .test_model_integration import *
